backend/requirements.txt
# ─── Web framework ────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# ─── Validation & config ──────────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1

# ─── AI / Gemini ──────────────────────────────────────────────────────────────
google-generativeai==0.8.3

# ─── Embeddings (local, no API key needed) ────────────────────────────────────
sentence-transformers==3.3.1
torch==2.5.1
transformers==4.47.1

# ─── Vector store ─────────────────────────────────────────────────────────────
chromadb==0.6.3

# ─── Document processing ──────────────────────────────────────────────────────
pypdf==5.1.0
python-docx==1.1.2

# ─── Data processing ──────────────────────────────────────────────────────────
pandas==2.2.3
numpy==1.26.4

# ─── HTTP client ──────────────────────────────────────────────────────────────
httpx==0.28.1

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# ─── Code quality ─────────────────────────────────────────────────────────────
ruff==0.8.4
